chore(retranspile): 删除 ai-config unit-variant sed——Plan 396 §2.5 根治后变 no-op
auto-lang plan-fix/396-unit-variant：parser 对三段限定 unit-variant 模式
（module.Type.UnitVariant，无载荷括号）剥模块段并转 TagCover，a2r 直接
产出 Value::Nil（与带参变体 Value::Str(s) 一致）。

- retranspile.sh 的 loader.rs sed（Nil/Null/Void ×3 unit 臂）实证 no-op 后删除
- rust/src 三文件追平当前 a2r：len() 强转与 int 局部 i32→i64
  （产物此前停在 Plan 021 时代，156c6c8 未覆盖 ai-config）
- retranspile.sh check：cargo check 0 错

+++ crates/ai-config/rust/src/loader.rs
@@ -145,7 +145,7 @@ impl TierRouting {
         return TierRouting { entries: std::collections::HashMap::new(), tier_names: vec![] };
     }
     pub fn is_empty(&self) -> bool {
-        return (self.tier_names.len() as i32) == 0;
+        return (self.tier_names.len() as i64) == 0;
     }
     pub fn candidates(&self, tier: &str) -> Option<Vec<TierRouteCandidate>> {
         if self.entries.contains_key(tier) {
@@ -214,7 +214,7 @@ pub fn parse_client_config(content: &str) -> Result<ClientConfig, ConfigError> {
                             let names = provider_keys(client_node.clone());
                             
 
-                            if (names.len() as i32) == 0 {
+                            if (names.len() as i64) == 0 {
                                 return Err(ConfigError::Parse("no providers configured in client { } block".to_string()));
                             }
                             let dp = if s.default_provider.is_empty() {
@@ -276,7 +276,7 @@ pub fn parse_daemon_config(content: &str) -> Result<DaemonConfig, ConfigError> {
                     cfg.tier_routing = parse_tier_routing(daemon_node.clone());
                     
 
-                    if cfg.default_provider.is_empty() && (cfg.provider_names.len() as i32) > 0 {
+                    if cfg.default_provider.is_empty() && (cfg.provider_names.len() as i64) > 0 {
                         cfg.default_provider = cfg.provider_names[0].clone();
                     }
                     if cfg.default_model.is_empty() {
@@ -285,7 +285,7 @@ pub fn parse_daemon_config(content: &str) -> Result<DaemonConfig, ConfigError> {
 
                             match cfg.providers.get(&cfg.default_provider) {
                                 Some(p) => {
-                                    if (p.models.len() as i32) > 0 {
+                                    if (p.models.len() as i64) > 0 {
                                         cfg.default_model = p.models[0].clone().id.to_string();
                                     }
                                 },
@@ -409,7 +409,7 @@ fn parse_tier_routing(node: Node) -> TierRouting {
 
                         } else {
                             let cands = parse_candidates(tier_val.clone());
-                            if (cands.len() as i32) > 0 {
+                            if (cands.len() as i64) > 0 {
